Category validation added to product pagination tests. TODO: Make repository. Make stubs and mocks of the validation functions and use them in the validation tests when not the SUT.

+++ products/src/test/setup.ts
@@ -38,9 +38,11 @@ beforeEach(async () => {
   const collections = await client.getDatabase()!.collections();
 
   for (let collection of collections) {
-    await collection.deleteMany({});
-    // recreate the index after cleaning our slate
-    await collection.reIndex();
+    if (collection.collectionName !== "categories") {
+      await collection.deleteMany({});
+      // recreate the index after cleaning our slate
+      await collection.reIndex();
+    }
   }
 });
